pyproject.toml
[project]
name = "clickup-cli"
version = "0.1.0"
description = "CLI tool for managing ClickUp tasks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.scripts]
cl = "clickup_cli.cli:cli"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"
